+++ NEWS.md
@@ -1,5 +1,12 @@
 # tarchetypes 0.2.0.9000
 
+## Bug fixes
+
+* Make the `*_raw()` target factories process `command` the same way whether it is an expression or ordinary language object.
+
+## New features
+
+* Add `tar_rep_map()` and `tar_rep_map_raw()` to perform batched computation downstream of `tar_rep()` (#50).
 
 
 # tarchetypes 0.2.0